requirements.txt
websockets>=12.0
aiohttp>=3.9.0
watchdog>=4.0.0
Pillow>=10.2.0
colorama>=0.4.6
pyyaml>=6.0.1
rich>=13.7.0
aiofiles>=24.1.0